Time: 17 ms (55.12%), Space: 9.7 MB (42.24%) - LeetHub

+++ 1553-minimum-number-of-days-to-eat-n-oranges/1553-minimum-number-of-days-to-eat-n-oranges.cpp
@@ -1,19 +1,19 @@
 class Solution {
 public:
-    unordered_map<long long int,long long int> mp;
-    long long int solve(long long int n){
+    unordered_map<int,int> mp;
+    int solve(int n){
         if(n<=0) return 0;
         if(n==1) return 1;
         
         if(mp.count(n) > 0) return mp[n];
         
-        long long int op3=INT_MAX;
+        int op3=INT_MAX;
         //op1 = 1 + solve(n-1);//no need of eating one day one orange
         //but to make n divisible by 2 nd 3 we will use n%2,3;
         // n/3 = n - 2n/3
         //n%3 -> 0,1,2 and n%2 -> 0,1
         op3 = 1 + (n%3) + solve(n/3);
-        long long int op2 = 1 + (n%2) + solve(n/2);
+        int op2 = 1 + (n%2) + solve(n/2);
         
         return mp[n] = min(op2,op3);
     }
